requirements.txt
flowserv-core>=0.2.0
matplotlib
